pyproject.toml
[tool.poetry]
name = "api-yamdb"
version = "0.1.0"
description = ""
authors = ["David Saidov <delightxxls@gmail.com>"]
readme = "README.md"
packages = [{include = "api_yamdb"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "2.26.0"
django = "3.2"
djangorestframework = "3.12.4"
pyjwt = "2.1.0"
pytest = "6.2.4"
pytest-django = "4.4.0"
pytest-pythonpath = "0.7.3"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.6"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
